pyproject.toml
[project]
name = "gatekeeper"
version = "0.1.0"
description = "Shared authentication service and library"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.0",
    "apsw>=3.45",
    "itsdangerous>=2.1",
]

[project.optional-dependencies]
http = ["httpx>=0.27"]
ldap = ["python-ldap>=3.4"]
dev = [
    "ruff",
    "ty",
    "pytest",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["gatekeeper*", "gatekeeper_client*"]

[tool.setuptools.package-data]
gatekeeper = ["templates/**/*.html", "static/**/*"]

[tool.ruff]
target-version = "py314"
line-length = 100
lint.select = ["E", "F", "I", "UP", "B"]
lint.ignore = ["E501"]

[tool.ty.environment]
python-version = "3.14"
